Modificacion de tipo de variables float por double y comenzamos con la clase main

+++ src/controller/FlowerShopController.java
@@ -22,17 +22,17 @@ public class FlowerShopController {
 		this.name = name;
 	}
 	
-	public void createTree(int code, String name, float price, int stock,  float height) throws Exception {
+	public void createTree(int code, String name, double price, int stock,  double height) throws Exception {
 		Tree tree = new Tree(code, name, price, stock, height);
 		Repository.addItems(tree);	
 	}
 	
-	public void createFlower(int code, String name, float price, int stock,  String color) throws Exception {
+	public void createFlower(int code, String name, double price, int stock,  String color) throws Exception {
 		Flower flower = new Flower(code, name, price, stock, color);
 		Repository.addItems(flower);	
 	}
 	
-	public void createDecoration(int code, String name, float price, int stock,  String material) throws Exception {
+	public void createDecoration(int code, String name, double price, int stock,  String material) throws Exception {
 		Decoration deco = new Decoration(code, name, price, stock, material);
 		Repository.addItems(deco);	
 	}
